+++ src/pages/index.tsx
@@ -1,19 +1,24 @@
-import { useRouter } from 'next/router';
+import React from 'react';
+import Layout from '../components/Layout';
+import { Card, Space } from 'antd';
 
-export default function Main() {
-  const router = useRouter();
-
-  const handleChangeRouter = () => {
-    router.push('/home');
-  };
-
+export default function Home() {
   return (
     <>
-      <div>test</div>
-      <button className="cypressTest" onClick={handleChangeRouter}>
-        {' '}
-        cypress{' '}
-      </button>
+      <Layout>
+        <Space direction="vertical" size={16}>
+          <Card title="Default size card" extra={<a href="#">More</a>} style={{ width: 300 }}>
+            <p>Card content</p>
+            <p>Card content</p>
+            <p>Card content</p>
+          </Card>
+          <Card size="small" title="Small size card" extra={<a href="#">More</a>} style={{ width: 300 }}>
+            <p>Card content</p>
+            <p>Card content</p>
+            <p>Card content</p>
+          </Card>
+        </Space>
+      </Layout>
     </>
   );
 }
